requirements.txt
asgiref==3.6.0
Django==4.2.1
django-debug-toolbar==4.1.0
django-mysql==4.9.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
